+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - The `Spinner` component.
 - The `Loader` component.
+- The `Panel` component.
 
 ### Changed
 - Renamed animations to be consistent with naming style throughout the rest of the component library e.g., `chqSpin`.